some bugfix and new XML feature (#206)
* some bugfix and new XML feature

* add JSON to drlm-error-send

* https support in xml/json send error

* update realease notes

+++ usr/share/drlm/lib/install-functions.sh
@@ -527,9 +527,11 @@ function remove_client_scripts () {
 
 function tunning_rear () {
 
-  local REAR_VERSION=$(rear -V | awk '{print $2}')
+  local REAR_VERSION=$($SUDO rear -V | awk '{print $2}')
   if [ "$REAR_VERSION" == "2.00" ]; then
-    $SUDO sed -i 's%https://$DRLM_SERVER/clients/$DRLM_ID%https://$DRLM_SERVER/clients/$DRLM_ID/config/default%g' /usr/share/rear/lib/drlm-functions.sh
+    if ! grep -q 'https://$DRLM_SERVER/clients/$DRLM_ID/config/default' /usr/share/rear/lib/drlm-functions.sh; then
+      $SUDO sed -i 's%https://$DRLM_SERVER/clients/$DRLM_ID%https://$DRLM_SERVER/clients/$DRLM_ID/config/default%g' /usr/share/rear/lib/drlm-functions.sh
+    fi
   fi
 
   # remove cURL verbose to avoid infinite lines of Debug in some cURL versions
@@ -539,11 +541,13 @@ function tunning_rear () {
   
   # Solve SELinux autorelabel after recover (RSYNC)
   if [ -f "/usr/share/rear/restore/default/500_selinux_autorelabel.sh" ]; then
-    $SUDO sed -i '/^touch \$TARGET_FS_ROOT\/\.autorelabel/i rm -rf \$TARGET_FS_ROOT\/\.autorelabel' /usr/share/rear/restore/default/500_selinux_autorelabel.sh
+    if ! grep -v '^\s*$\|^\s*\#' /usr/share/rear/restore/default/500_selinux_autorelabel.sh | grep -q 'rm -rf $TARGET_FS_ROOT/.autorelabel'; then
+      $SUDO sed -i '/^touch \$TARGET_FS_ROOT\/\.autorelabel/i rm -rf \$TARGET_FS_ROOT\/\.autorelabel' /usr/share/rear/restore/default/500_selinux_autorelabel.sh
+    fi
   fi
 
   # remove rear cron file
-  if [ -f "/etc/cron.d/rear" ]; then
+  if $SUDO test -f "/etc/cron.d/rear"; then
     $SUDO rm -rf /etc/cron.d/rear
   fi
 }
